requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
langgraph>=0.0.24
langchain-core>=0.0.27
langchain-tavily>=0.0.1
pydantic>=2.0.0
python-dotenv>=1.0.0